02-rag-get-started/requirements.txt
langchain
langchain-core
langchain-text-splitters
langchain[openai,azure]
chromadb
gradio
pypdf
pydantic
python-dotenv